+++ build.gradle
@@ -1,27 +1,14 @@
+buildscript {
+    repositories {
+        mavenCentral()
+    }
+}
+
 plugins {
-    id 'java'
-    id 'org.springframework.boot' version '3.0.2'
-    id 'io.spring.dependency-management' version '1.1.0'
 }
 
-group = 'de.tum.cit.artemis.push'
-version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '17'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-
-    implementation "com.google.firebase:firebase-admin:9.1.1"
-
-    implementation 'com.eatthepath:pushy:0.15.2'
-
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-}
-
-tasks.named('test') {
-    useJUnitPlatform()
+allprojects {
+    repositories {
+        mavenCentral()
+    }
 }
