+++ common/src/main/java/de/tum/cit/artemis/push/artemispushnotificationrelay/common/SendService.java
@@ -1,0 +1,8 @@
+package de.tum.cit.artemis.push.artemispushnotificationrelay.common;
+
+import org.springframework.http.ResponseEntity;
+
+public interface SendService<T> {
+
+    ResponseEntity<Void> send(T request);
+}
