+++ firebase/build.gradle
@@ -1,0 +1,25 @@
+plugins {
+    id 'java'
+    id 'org.springframework.boot'
+    id 'io.spring.dependency-management'
+}
+
+group = 'de.tum.cit.artemis.push.artemispushnotificationrelay'
+version = '0.0.1-SNAPSHOT'
+sourceCompatibility = '17'
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    implementation 'org.springframework.boot:spring-boot-starter'
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+
+    implementation "com.google.firebase:firebase-admin:9.1.1"
+    implementation(project(":common"))
+}
+
+tasks.named('test') {
+    useJUnitPlatform()
+}
