+++ settings.gradle
@@ -1,1 +1,18 @@
+pluginManagement {
+    repositories {
+        mavenCentral()
+    }
+
+    plugins {
+        id 'java' apply false
+        id 'org.springframework.boot' version '3.0.2' apply false
+        id 'io.spring.dependency-management' version '1.1.0' apply false
+    }
+}
+
+include("common")
+include("apns")
+include("firebase")
+include("hermes")
+
 rootProject.name = 'artemis-push-notification-relay'
