--- src/main/java/de/tum/cit/artemis/push/artemispushnotificationrelay/DeviceType.java
@@ -1,6 +1,0 @@
-package de.tum.cit.artemis.push.artemispushnotificationrelay;
-
-public enum DeviceType {
-    FIREBASE,
-    APNS
-}
