--- src/main/java/de/tum/cit/artemis/push/artemispushnotificationrelay/FirebaseSendPushNotificationsRequest.java
@@ -1,8 +1,0 @@
-package de.tum.cit.artemis.push.artemispushnotificationrelay;
-
-import java.util.List;
-
-record FirebaseSendPushNotificationsRequest(
-    List<NotificationRequest> notificationRequest
-) {
-}
