--- src/main/java/de/tum/cit/artemis/push/artemispushnotificationrelay/SendService.java
@@ -1,10 +1,0 @@
-package de.tum.cit.artemis.push.artemispushnotificationrelay;
-
-import org.springframework.http.ResponseEntity;
-
-import java.util.List;
-
-interface SendService<T> {
-
-    ResponseEntity<Void> send(T request);
-}
