--- src/test/java/de/tum/cit/artemis/push/artemispushnotificationrelay/ArtemisPushNotificationRelayApplicationTests.java
@@ -1,13 +1,0 @@
-package de.tum.cit.artemis.push.artemispushnotificationrelay;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ArtemisPushNotificationRelayApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
